check_runner: Allow to overwrite timeout in result_set
> timeout = timeout or check.timeout or CHECK_TIMEOUT

Timeout for checks was based on 3 different values.
a) argument `timeout` which was passed from command line --timeout
b) timeout specified in check in result_set
c) predefined default value `CHECK_TIMEOUT`

The bug was that value of argument was changed in the function
and there is nothing like pass by value in python.

Right timeout was used for the 1st check but not for others.